+++ Package.swift
@@ -3,7 +3,7 @@ import PackageDescription
 let package = Package(
     name: "ParselyAnalytics",
     platforms: [
-        .iOS(.v12)
+        .iOS(.v13)
     ],
     products: [
         .library(